feat: inform user about offending restrictOnDelete records
Previously when attempting to delete e.g. Positions on which Figures
dependended (which would be prevented by restrictOnDelete foreign key
restraints), we warned the user that Figures dependended on the Position
and to delete the figures first, but didn't actually tell the user what
the figures where.
We now tells the user the names of the offending figures (or compound
figures when deleting figures). Still not an ideal UX because flash
message disappears, but better than before.

And in the process I added a config setting for the maximum number of
offending records to display, and renamed "defaults.php" to "misc.php"
and will use misc.php for a dump of miscellaneous config settings.