refactor(seed): specify CompoundFigureFigures from/to positions
Context: I'm switching to overloaded figure names, e.g. calling multiple
figures "Enchufa", and relying on the combination of name, from
position, and to position to uniquely identify a figure.

As a consequence we must now specify a figure's name *and* from/to
positions when seeding a CompoundFigure's CompoundFigureFigures, instead
of just specifying the figure name. This commit updates
CompoundFigureSeeder accordingly.