wip: add Graphviz node-link graph of figures and positions
This commit adds a Bash script casinograph.bash for generating a
node-link diagram representation of figures and positions (uses Graphviz
and saves as an SVG to public/img), and supporting code in
PositionController to regenerate the image when Positions change.
WIP for various reasons to be addressed shortly; commiting for now.