refactor: regenerate casino graph on visits to Home
Instead of regenerating the graph whenever models change in controllers,
why not just regenerate it on visits to Home using current database?
I mean, it takes ~50ms (okay, will be less on VPS, but probably not
orders of magnitude more), we can afford it---it's behind a rate limiter
anyway as a precaution against overwhelming the server.

The motivation for doing so, besides simplicity, is that I'd like to
randomly choose only one self-referencing edge for each node on each
request, to add variety, given that Graphviz seems to really struggle to
display multiple self-referencing nodes.