refactor: move compound figure integrity check to validator
Context: when updating a Figure's to/from position, we must check that a
new position will not corrupt the to/from position integrity in the
figure sequence of any dependent CompoundFigure.

I had previously done this in the update method of Figure controller,
when the appropriate place is during validation. The new validtion code
is added in this commit; I'll clean up the controller code when moving
controller code to dedicated services.