feat: implement RegenerateCasinoGraph on a per-user basis.
Previously an SVG digraph was generated by concatenating *all* queries
in the database. Now every user gets their own "sandboxed" graph with
only that user's positions and figures.

Summary: moved generation of DOT file to PHP and did away with
casinograph.bash, and updated queries and logic in RegenerateCasinoGraph
to work on a per-user basis. Added a few relevant config values. Also
now we pass the public path to SVG to frontend via a prop in Home.vue.