refactor: use casino user for public-facing resources
Previously: I had a special set of records with NULL user_id, and used
those for null-facing resources. It was a lot of headache to handle the
special case of null user_id and also to maintain a separate copy of
records for null users.

This commit does away with all of this null user trouble and just uses
a root `casino` user's resources (in practice Elijan's) as the
public-facing resources.

Changes:
- user_id cannot be null in migrations
- Define config value `config.constants.user_ids.casino`
- Create seeder in local and prod for `casino` user
- Update seeding routines to replace null user ids with
  `config.constants.user_ids.casino`
- Update models functions to return records of
  `config.constants.user_ids.casino` when Auth::id() is null.
- Update Casinograph generation function to return records of
  `config.constants.user_ids.casino` when Auth::id() is null.