feat: allow filtering Figures by from/to Position
In FigureTabPanel (so in Figures/Index, as far as the user is concerned)
allows filtering Figures by from/to Position (previously one could
filter only by FigureFamily).

The commit includes a nontrivial UI refactor: filters in FigureTabPanel
are now delegated to a Disclosure opened/closed by a "More filters"
button (to avoid cluttering the FigureTabPanel with what was becoming a
large number of filter elements).