feat: implement SVG position graphs in Positions/Show
Adds frontend and backend code for creating an SVG graph with Graphviz
for every position (showing the position's incoming and outgoing figures
with their corresponding from/to positions) and displaying this graph in
Positions/Show.vue.

The graph generation procedure is basically analogous to generating the
main CasinoGraph SVG shown on the Home page, with the extra nuance that
a separate graph has to be generated for every position, instead of a
single main CasinoGraph.

Also adds a custom session handler, MyFileSessionHandler, currently
empty and just a copy of default FileSessionHandler which will be used
to clean up a users' position graph SVGs from the file system when their
session ends (to avoid overcluttering the file system with SVGs).

todo: in `destroy` method of MyFileSessionHandler (currently a verbatim
copy of default FileSessionHandler, delete a user's position graph SVGs
from file system when the user's session ends).