revert: dispatch RegenerateSVG jobs from `show` methods
Reverts 7a43e26, which switched to dispatching RegenerateSVG jobs on
store/update. This doesn't work well for interconnected
resources—consider the situation where, say, you update the name of one
of a position's outgoing figures. Well, the position's SVG won't update
to reflect the new figure name! By far the simplest way to handle the
SVG situation is to just regenerate SVG graphs on `show`, and this seems
a light enough load that I can get away with it fine.