refactor: move Graphviz code to CasinoGraphService
Moves code related to generating the CasinoGraph SVG to a dedicated
CasinoGraphService, and jas the RegenerateCasinoGraph Job simply call
this service in an API-like fashion.

Motivation: separation of concerns, and looking ahead to using a
CasinoGraph SVG to debug SCC, so it will be helpful to have a separate
API-like function to call to generate graphs.