feat: only build FigureSequence from a SCC
Context: You only want to construct a figure sequence out of figures
that do not end at a "dead-end" position, i.e. a position with no
outgoing figures. In graph speak: you only want to construct the figure
sequence out of figures in a strongly connected component of the
CasinoGraph. This ensures the figure sequence can never dead-end.

This commit modifies FigureSequenceService to only build a figure
sequence from the largest SCC in the user's CasinoGraph, solving the
dead-end problem for figure sequences.